Split StaticDisplayConfig into its own files
