+++ debian/control
@@ -54,7 +54,8 @@ Build-Depends: cmake,
                libxcb-xfixes0-dev,
                libxcb-render0-dev,
                libxcb-composite0-dev,
-               libxcursor-dev
+               libxcursor-dev,
+               libyaml-cpp-dev
 Standards-Version: 3.9.4
 Homepage: https://launchpad.net/mir
 # If you aren't a member of ~mir-team but need to upload packaging changes,
